refactor/ rotas de clinica e paciente - adicionado o método PUT

+++ app/api/clinics/clinic_route.py
@@ -1,6 +1,6 @@
 from typing import List
 
-from fastapi import APIRouter, status
+from fastapi import APIRouter, HTTPException, status
 
 from app.models.clinic_models import Clinic, ClinicAdminResponse
 from app.service.clinic_service import ClinicService
@@ -46,8 +46,28 @@ def get_clinic_by_id(id: str):
     """
     Busca um determinado documento dentro da Collection por id
     """
+    if not id:
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail='Nenhum id foi enviado.',
+            )
     return clinic_service.get_clinic_by_id()
 
+
+@clinic_route.put(
+    '/id/{id}',
+    status_code=status.HTTP_200_OK,
+    response_model=ClinicAdminResponse,
+    response_model_by_alias=False
+)
+def update_clinic_by_id(id: str, data: Clinic):
+    if not id:
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail='Nenhum id foi enviado.',
+            )
+    return clinic_service.update_clinic_by_id(id, data)
+    
 
 @clinic_route.delete(
     '/{id}',
@@ -57,4 +77,9 @@ def delete_clinic_by_id(id: str):
     """
     Deleta uma clínica por id (uuid4)
     """
+    if not id:
+            raise HTTPException(
+                status_code=status.HTTP_400_BAD_REQUEST,
+                detail='Nenhum id foi enviado.',
+            )
     clinic_service.delete_clinic_by_id(id)
